refactor: split Project Subcription Alert scene's protocols into separated folder and files
