+++ Pharus/Scenes/Alerts/Project Subcription Alert/Protocols/ProjectSubcriptionAlertFlow.swift
@@ -1,0 +1,12 @@
+//
+//  ProjectSubcriptionAlertFlow.swift
+//  Pharus
+//
+//  Created by Victor Colen on 17/06/22.
+//
+
+import Foundation
+
+protocol ProjectSubcriptionAlertFlow {
+    func closeModal()
+}
