+++ Pharus/Scenes/Alerts/Project Subcription Alert/Protocols/ProjectSubcriptionAlertViewDelegate.swift
@@ -1,0 +1,14 @@
+//
+//  ProjectSubcriptionAlertViewDelegate.swift
+//  Pharus
+//
+//  Created by Victor Colen on 17/06/22.
+//
+
+import Foundation
+
+protocol ProjectSubcriptionAlertViewDelegate: AnyObject {
+    func closeButtonTapped()
+    func primaryButtonTapped()
+    func secondaryButtonTapped()
+}
