+++ Pharus/Scenes/Alerts/Project Subcription Alert/Protocols/ProjectSubcriptionPresenterProtocol.swift
@@ -1,0 +1,13 @@
+//
+//  ProjectSubcriptionPresenterProtocol.swift
+//  Pharus
+//
+//  Created by Victor Colen on 17/06/22.
+//
+
+import Foundation
+
+protocol ProjectSubcriptionPresenterProtocol {
+    func closeModal()
+    func subscribeToProject()
+}
